+++ cli/src/instructions/admin/update_base_fee.rs
@@ -1,5 +1,4 @@
 use crate::*;
-use anchor_client::solana_sdk::transaction::Transaction;
 
 #[derive(Debug, Parser)]
 pub struct UpdateBaseFeeParams {